可以发送按键消息了
，但是不知道对不对，没测试，现在服务端运行不了

+++ TankCraftServer/听君语/Text_RenderComponent.cpp
@@ -17,7 +17,7 @@ Xn::Text_RenderComponent::Text_RenderComponent(const Vector2& start_pos,
       &text_format));
 
   Ìý¾ýÓï::Get().GetRenderManager()->dwrite_factory_->CreateTextLayout(
-      text.data(), text.size(), text_format.Get(), WH.asWH.width,
+      text.data(), (uint)text.size(), text_format.Get(), WH.asWH.width,
       WH.asWH.height, &text_);
 }
 
@@ -43,13 +43,13 @@ void Xn::Text_RenderComponent::SetText(const std::wstring& text) {
   auto height = text_->GetMaxHeight();
 
   Ìý¾ýÓï::Get().GetRenderManager()->dwrite_factory_->CreateTextLayout(
-      text.data(), text.size(), text_format.Get(), width, height, &text_);
+      text.data(), (uint)text.size(), text_format.Get(), width, height, &text_);
 }
 void Xn::Text_RenderComponent::SetWH(const Vector2& WH) {
   text_->SetMaxWidth(WH.asWH.width);
   text_->SetMaxHeight(WH.asWH.height);
 }
-void Xn::Text_RenderComponent::SetFontSize(const uint& size) {
+void Xn::Text_RenderComponent::SetFontSize(const Float& size) {
   DWRITE_TEXT_RANGE text_range = {.startPosition = 0, .length = 256};
   text_->SetFontSize(size, text_range);
 }
